Created account service and repo. Moved logic for login,register and token from user to account service.
