+++ BackendDotnet/Keeper.Main/Controllers/AccountController.cs
@@ -1,6 +1,7 @@
 ﻿using Keeper.Common.Response;
 using Keeper.Common.ViewModels;
 using Keeper.Services.Interfaces;
+using Keeper.Services.Services.Interfaces;
 using Microsoft.AspNetCore.Mvc;
 
 
@@ -9,21 +10,21 @@ namespace Keeper.Main.Controllers
     [Route("api/[controller]")]
     [ApiController]
     public class AccountController : ControllerBase
-    {
-        private readonly IUserService _userService;
-        public AccountController(IUserService userService)
+    {  
+        private readonly IAccountService _accountService;
+        public AccountController(IAccountService accountService)
         {
-            _userService = userService;
+            _accountService= accountService;
         }
         [HttpPost("Register")]
         public async Task<ResponseModel<string>> Register(RegisterVM register)
         {
-            return await _userService.RegisterAsync(register);
+            return await _accountService.RegisterAsync(register);
         }
         [HttpPost("Login")]
         public async Task<ResponseModel<TokenModel>> Login(LoginVM loginVM)
         {
-            return await _userService.LoginAsync(loginVM);
+            return await _accountService.LoginAsync(loginVM);
         }
     }
 }
