+++ BackendDotnet/Keeper.Repos/Repositories/Interfaces/IAccountRepo.cs
@@ -1,0 +1,9 @@
+﻿using Keeper.Context.Model;
+
+namespace Keeper.Repos.Repositories.Interfaces
+{
+    public interface IAccountRepo
+    {
+        Task<bool> RegisterAsync(UserModel user);
+    }
+}
