+++ BackendDotnet/Keeper.Repos/Repositories/Interfaces/IUserRepo.cs
@@ -4,7 +4,6 @@ namespace Keeper.Repos.Interfaces
     public interface IUserRepo
     {
         Task<IEnumerable<UserModel>> GetAllAsync();
-        Task<bool> RegisterAsync(UserModel user);
         Task<UserModel> GetByEmailAsync(string email);
         Task<UserModel> GetByIdAsync(Guid id);
     }
