+++ BackendDotnet/Keeper.Services/Services/Interfaces/IAccountService.cs
@@ -1,0 +1,11 @@
+﻿using Keeper.Common.Response;
+using Keeper.Common.ViewModels;
+
+namespace Keeper.Services.Services.Interfaces
+{
+    public interface IAccountService
+    {
+        Task<ResponseModel<string>> RegisterAsync(RegisterVM register);
+        Task<ResponseModel<TokenModel>> LoginAsync(LoginVM loginVM);
+    }
+}
